Allow overriding the APIS_ENTITIES setting
The APIS_ENTITIES are created using the values that are stored in
the `entity_settings` attribute of entities. In some cases it might
make sense to be able to define those in the global settings file.
This commit changes the code so it checks for the existence of the
APIS_ENTITIES setting and only uses the `entity_settings` attribute
if the global settings is empty.

This commit also changes the default_cols in the get_table view
to respect the default columns in the settings. Given that in
apis_core.apis_entities.models settings from models and settings are
harmonized this should allow to set it in models.py as well as in
settings.