fix: apis_entities: get filter fields from APIS_ENTITIES settings
As discussed in acdh-oeaw/apis-ontologies#9 the entity configuration should be
overridable using settings. There is also logic to set store `entity_settings`
attribute of an entity in the settings.APIS_ENTITIES, so it seems
settings.APIS_ENTITIES is the source of truth for those settings.
This change fixes the `set_sort_filters` method to retrieve the list of filters
from settings.APIS_ENTITIES instead from the `entity_settings` attribute. The
commit also refactores the code a bit.