apis_entities: refactor get_all_entity_classes
This commit refactors the get_all_entity_classes() method. Instead of loading a
hardcoded module we simply get a list of all models and check wether they are
children of TempEntityClass (or in the future of another class that is the root
of all entities). This way we can also loose the inspect import.