Drop the infos package
The infos package is independend of the apis codebase. It is a totally separate
Django application, that can and should reside in its own git repository as a
separate package. This git repository is now on
https://github.com/acdh-oeaw/apis-infos - therefore we can remove the app from
this repo.
This not only deletes the files, but also removes the `infos` app from the
shipped base settings, removes the `infos` paths from the url file and removes
the app from the pyprojec.toml file.

Closes: #51