Revert "apis_entities: refactor get_all_entity_classes"
This reverts commit 62565e423719a726660360cc43f72e12cdaa5520.
The commit introduced a change that broke migrations. Given that the
method is part of a model, but accesses a list of all models, this leads
to the situation that not all models are loaded yet when the model is
needed. The class will be removed anyway, so we can revert the change
for now an look for a better solution later.