Clean up settings file for CI tests
This old settings file contained various settings for CI tests on gitlab, but
it hasn't been used anymore. We reuse the settings file and remove the old
entries. Ideally the tests would be able to run with the base settings file of
apis, but given some peculiarities of the default settings we have to override
some of the them to be able to run the tests (but maybe, some day ...?)