docs: automatically generate module documentation
Up until now, the documentation of the modules was controlled by the
`source/modules` folder, which contained *.rst files for a couple of
modules, but only those got autogenerated documentation.
sphinx provides the `sphinx-apidoc` tool which can generate the *rst
files for us. So we drop the old `source/modules/*.rst` files and let
`sphinx-apidoc` generate them on the fly from now.
This commit adds this command to the Makefile, it adds the
source/modules folder to the .gitignore, so we don't accidently commit
those autogenerated files again, and it replaces the old menu entry in
the documentation with a link to the generated docs.
Last but not least, we fix the excludes directive in the sphinx conf.py
to ignore the migrations modules.