apis_entities: fix merging two entities
The RDFParser needs the correct app_label_entities set, to find the
ContentType. We could also fix the RDFParser directly, but at the
current state this could have unknown consequences.
In the merge method of the TempEntityClass it seems that the lookup
of the ContentType also works withtout setting the app_label.