Drop the browsing module
In a JFX discussion it wasn't really clear what the module does. In any
case, it is not a part of apis-core and should therefore be moved to a
separate git repository. Users can include it from there in their apis
setup if needed.

The apis-browsing modules now lives in
https://github.com/acdh-oeaw/apis-browsing

A couple of template files, views and templatetags were copied to the
`apis_metainfo` module:
`browsing/browsing_utils.py` is now
`apis_core/apis_metainfo/browsingviews.py`. This module was copied,
because there are some views in `apis_metainfo` that inherit from the
browsingviews. They should probably be reviewed at some point, but for
now simply copying the views was the less intrusive option. This also
means that we have to copy the following templates:
`create_base_template.html`, `generic_create.html`, `generic_list.html`,
`class_definition.html`, `column_selector.html` and the templatetags
from `browsing_extras.py`