feat: implement duplication of RootObject children
This commit adds a method `duplicate` to the RootObject class. The
method creates a copy of self and tries to copy some of the relations
too. This does not copy any triples for now, given that turned out to be
more complicated than expected.
The commit also introduces a `GenericEntitiesDuplicateView` in
apis_entities.edit_generic, a `get_duplicate_url` in
`apis_entities.models` and an url entry for that view. Last but not
least it adds a duplicate link to the `detail_views/detail_generic.html`
template.