fix: Set the default_auto_field in all the apps
Maintaining the historical behavior, the default value for
DEFAULT_AUTO_FIELD is AutoField. We are setting this value on a per app
basis to avoid unwanted migrations for users.
Starting with 3.2 new apps are generated with
AppConfig.default_auto_field set to BigAutoField.
See https://docs.djangoproject.com/en/3.2/releases/3.2/#customizing-type-of-auto-created-primary-keys