style: rename `helper_functions` to `utils`
The idea of the folder `helper_function` is to provide reusable code
that can be used by different components of apis. This is not restricted
to functions, but it can also be classes, enums, views, mixins,
fields... a more fitting name is be utils, like it is also used in
Django itself or DRF.
BREAKING CHANGE: external projects using functionality from
`helper_functions` will have to update their imports.
Closes: #131