fix(apis_entities): fix transferral of Uris in merge method
The existing for loop that should update the root_object of the Uris
sets the attribute `entity` which is not a field in the Uri model. It
simply gets lost when saving. The field is called `root_object`. We fix
this and replace the for loop with a filter update, which would have
shown the mistake in the first place.