feat(apis_entities): refactor code and make custom filters possible
This commit refactors the `set_sort_filters` filters method on the
GenericEntityListFilter and renames it to `set_and_sort_filters`
It changes how the APIS_ENTITIES.list_filters setting is parsed:
This setting has to be a dict now. The keys of the dict list the various
filters - either ones that refer to existing filters or new ones. The
values of the dict are dicts describing the filter.
In addition, we add a `list_filters_exclude` setting, which is a simple
list of string with names of filters that should be excluded. There is
the special word `__defaults__` which means that the default filters are
excluded, meaning only the filters in the `list_filters` setting will be
part of that filterset.
BREAKING CHANGE: this changes the format of the `list_filters` setting
of the APIS_ENTITIES setting to a pure dict.

Closes: #144