refactor!: move filtermethods from apis_entities to utils
The apis_entities.filters.GenericEntityListFilter contained some filter
methods (as defined in
https://django-filter.readthedocs.io/en/main/ref/filters.html#method)
and a helper method. Both could be useful to other modules. Therefore
those methods are being moved out of the GenericEntityListFilter to
`utils.filtermethods`.
In the GenericEntityListFilter the previously defined `name` filter is
not being used anymore.
The function names in the `apis_core.utils.filtermethods` module were
renamed and harmonized, as to not end in `_method` or `_filter`.
BREAKING CHANGE: if the GenericEntityListFilter was used with the
`method` set to one of the included methods, the `method` attribute
should now point to `apis_core.utils.filtermethods...`

Closes: #152