fix: harden and enhance Uri model
This commit introduces an UriManager for the UriModel. This way we can
normalize the 'uri' field when doing a `Uri.objects.get_or_create()`
instead of creating a new Uri object with a non normalized Uri.

This commit also adds a `clean()` method to the Uri model, that
normalizes the uri and checks if a root_object exists and creates it if
it doesn't. It raises exception if the `root_object` can not be created.