feat: Implement form, view and route for adding URIs
This commit introduces a form that gets or creates an URI. It creates a
view and the corresponding route for the view as well as a template for
the view.
There is also a subtemplate `gnd_autocomplete_form.html` which is
included in the URI get_or_create form - it provides an autocomplete
form that searches for entities in GND and on select adds the
corresponding URI to the input field.