fix(apis_metainfo)!: rename templatetags
The browsing_extras templatetags were initially merged from the browsing
application to make apis-core not depend on browsing. The webpage
application also has browsing_extras templatetags and therefore the
identical names led to a `templates.E003` warning from Django.
This commit renames the templatetags module to `apis_metainfo_extras`.
It also moves the used templattag templates to the
templates/apis_metainfo/tags folder.
The commit also drops the `nav_menu` templatetag, because it turned out
that it was not used anywhere.

Closes: #163