feat(apis_metainfo)!: cleanup menu block in base template
The menu block contained a default `About` menu item, allowing the user
to add custom menu items. By default, that menu item is emtpy, so it
showed a dropdown menu without any menuitems.
This commit replaces this menu item with a `main-menu-extra` block. This
gives downstream projects more control over the menu, letting them add
not only menu items to an existing menu, but also additional menus.
The block around the existing dropdown menus was renamed: prior to this
it was called `custom_dropdown_menus` which is a bit misleading, given
it contains the default apis menu items. This block was renamed to
`main-menu` (like the id of the `ul` tag it is enclosed in).