feat(apis_entities,utils): introduce new tei renderer
This commit replaces the removed `apis_tei` module with a new renderer.
The renderer is using the template engine to create tei xml files. This
is done by introducing a default TeiRenderer, that uses templates to
return xml responses.
The existing EntityToTEI renderer now inherits from this renderer and
uses the entites type (person, place, institution) to set the template
name. A couple of defaults templates are shipped in the
`apis_entities/templates/apis_entities/tei` folder.