feat(apis_entities): let users override the serializer for renderers
This commit lets the users define a function in an entity for providing
a custom serializer for a specific renderer. The method in the model
that provides the custom serializer has to have a name, that corresponds
to the full path of the used renderer. So if the renderer
`apis_core.apis_entities.api_renderers.EntityToTEI` serializer is used,
the corresponding serializer should be returned by the method
`apis_core_apis_entities_api_renderers_EntityToTEI()`

This commit also introduces two new helper function in utils:
`get_module_path` and `get_python_safe_module_path`