feat(apis_metainfo): replace fundament with pure bootstrap
This commit replaces the used `fundament` stylesheet links and script
sources with pure bootstrap4 and feather. The bootstrap stylesheet link
was put in the `partials/bootstrap4_css.html` template partial, the
needed bootstrap4 javascript sources from bootstrap and feather are in
the `partials/bootstrap4_js.html` and `partials/feather_js.html`
templates.
The stylesheet `include` was put into a `styles` block.
To make the navbar reflect the navbar as it was with fundament, the
project-logo had to be hardcoded with a height of 36px. The mr-0 we
just introduced in c35d13b, to move the usermenu to the end of the
navbar, was replaced by setting the #main-menu to `mr-auto`.

Closes: #180