refactor(apis_metainfo): drop `browsingviews.py`
The `browsingviews.py` module was introduces as a temporary measure when
the `browsing` module ware removed from apis-core-rdf in 91464d0. The
module was only used in `views.py` for the Uri related views. This
commit drops the `browsingviews.py` module and adapts `views.py`
accordingly. It also updates the Uri related filter, form and table
classes.

Closes #138