feat(apis_entities): filter allowed related types to only allowed
In the entity edit view, filter related entity types to only those where
a relation property exists.

Currently, the relationed edit panel on the right shows all possible
entities, even if no relations can actually be created between the two.
This commit filters the entity types using the requested type, to limit
the relatable entities to only those for which a relationship can
actually be created.