fix(apis_relations): fix accidental newline after `if` in template
In dba667f we reformatted all templates automatically using djlint.
dlint added newlines before and after `if` statements, which leads to
broken javascript, if the `if` statement is used to generate a function
argument. This commit turns djlint off for in two places where such
arguments are created using `if` statements (the edit and delete buttons
of the relations form).

Closes: #236