refactor: delete unused files and code
The files in apis_entities/
* `autocomplete_light_registry.py`
* `custom_autocompletes.py`
are not being used anywhere

`apis_entities/custom_authentication.py` is not being used, but could be
used by downstream projects. It was moved to
`apis_core.utils.authentication.py`.

The files in apis_relations/
* `autocomplete_light_registry.py`
* `serializers.py`
are not being used anywhere

The file `apis_relations/forms.py` was only used in
`apis_relations/views.py` and in there the result of the function that
used it was assigned to a variable that was not used anywhere. The whole
code block is now deleted.