feat(utils): replace `caching` module with `helpers`
The `caching` module is mostly implemented using code obfuscation and it
reimplements functionalitis from core Python and from Django.
This commit introduces a couple of `utils.helpers` functions that use
the Pyton `functools.cache` to provide the same functionality.

For now we are not deleting the `caching` module completely, because
there are still projects using it.