fix: replace custom duplication logic with signal
For duplicating entities, there was custom logic to copy ManyToManyField
- which was tailored to TempTriples. This had the downside of simply
creating new objects by hand.

This commit introduces a `duplicate` method in the Triple model, which
implements the functionality of duplicating itself and also triggers
pre_- and post_duplicate signals. This has the advantage that instances
that point to the Triple (like References) can also be duplicated using
signals.