fix(deps): bump yaml to latest
> poetry add pyyaml@latest

This is a copy of a hotfix that was merged to main - see
https://github.com/acdh-oeaw/apis-core-rdf/commit/dff296b47d017b8b24b5b5345e0c0de59c5e63d5
and
https://stackoverflow.com/questions/76708329/docker-compose-no-longer-building-image-attributeerror-cython-sources

The commit to former main did bump pyyaml, but only to 6.0 which did not
contain the fix. 6.0.1 does, but poetrys version resolving mechanism
chose the latest 6.0 version, so 6.0.1 was set in poetry.lock.
But we should make it clear, that we need the 6.0.1 version, so we are
setting that in pyproject.toml.