fix: show reference button in detail view
There are projects that use references on relations. Therefore the
references button should also be listed in the reference table - this is
already the case in the edit view, but not yet in the detail view. This
commit hardcodes this button in the detail view. This is a hack, but
given that the whole Reference logic is rewritten, it does not make much
sense to refactor the whole codebase to make this table more
customizable.