feat: add comments to revisions
This commit makes use of the reversion `set_comment` functionality.
Using comments, we can store additional information in the revisions.
That way we store the information if an object got added, if it was
updated and which fields. If another object was merged into it, we store
the original name of the other object. If an entity was created by
duplicating it, we store from which object it was duplicated.