feat: introduce new templatetags
This commit introduces a core templatetag `shared_url` to access the
`SHARED_URL` setting (the templatetag defaults to `/static/`). It also
updates the base template to use this templatetag.

The commit also introduces an apis_entities templatetag
`entities_list_links` that returns a list of links to the list views of
all entities. It also updates the base template to use this templatetag.

Closes: #293