refactor(apis_entities): rewrite the relevant fields logic
The approach so far was to manually go through all the fields and have
an `exclude` list. It makes more sense to use `model_to_dict` from
django forms. If projects need more customization, they can override the
template.
The variable was also renamed to `relevant_fields` and we also pass the
field to the template, to be able to access the verbose_name.

Closes: #331