refactor: move column selector from form to table
This commit changes the way the `column_selector` is set up. Up until
now, it was a simple select input, that could only be used in
conjunction with an existing form. That meant, that the colums selector
had to be places at a position away from the actual table.
Now the column selector is its own form, therefore it can be places
whereever we want it. Also it is not a templatetag anymore, but a simple
partial template that has to be included with {% include ... %}.
The downside is, that we might now have multiple forms sending get
parameters - this is where the new "form_get_hidden.html" template comes
in. It adds existing get parameters to hidden fields in the form, that
way multiple forms don't get in each others way.

Closes: #173