fix: do not expect fields in EntitySerializer
We are moving away from expecting fields in our entities, therefore it
makes sense to make the EntitySerializer a bit more flexible. By default
it serializes all the fields of the entity anyway - so we can simply
remove `name`, `start_date` and `end_date` in this case. The `label_set`
is a leftover of the legacy setup with TempEntityClass, we no check for
its existence before adding the field.

Closes: #383