fix: cleanup entities detail template
The detail template references a couple of attributes that can not be
expected of entities anymore. In addition, due to the reimplementation
of the `relevant_fields` logic in 13986b8 the template does not have to
list attributes manually. Therefore we remove some attributes from the
listing and others are now surrounded by {% if %}{% endif %} statements.

Closes: #396