feat: introduce `get_member_for_entity` helper
This function can help override objects - we can use it in the code to
look for members with the name derived from a model name. This
way we can override classes and members in general in the code just by
providing the object, but without having to define them in the settings.