feat: implement dumpdata view and admin command
This feature provides ways to serialize the data of an APIS instance to
be able to import it in a freshly installed instance later on.

The main logic is implemented in the `utils.helper.datadump_serializer`
method. It serializes the important models to json - with natural
foreign keys.

There is both a view ('/api/dumpdata') and an admin command
('apisdumpdata') that output the serialized data as API Response or to
stdout.

Closes: #273