fix: cleanup generic edit template
The card-header and card-body divs are now being enclosed in blocks that
can be used to override them. The object_texts block was removed, as
there is no variable anymore in the context. All the if/else checks if
there is an instance are now removed, given that the createviews now
have their own template.