refactor(apis_vocabularies)!: remove apis_vocabularies app
This removes the `apis_vocabularies` app- the functionality is not being
used right now and it will be reindtroduced in #441.
The commit removes the app and all references to it from other apps.
We drop the whole `apis_vocabularies` Django app - but because there are
still references from other apps migrations to this apps migrations and
to clean up existing databases, we don't remove the whole module, but
still ship the migrations.

Closes: #514