refactor(apis_entities)!: use new generic app
This commit drops all the existing entities views and recreates them
using the `generic` app as a basis.

Additionaly, it adds a merge view and a merge form.

The commit deletes the following files, because they are not used
anymore:
* detail_generic.py, edit_generic.py: those are now handled by
  generic.views and apis_entities.views
* merge_views.py: replaced by a new merge view in apis_entities.views
* filters.py, tables.py: were only used by GenericListViewNew, which is
  now replaced by generic.views.List

BREAKING CHANGE: with the removal of `filters.py` and `tables.py` your
configuration in `APIS_ENTITIES` won't have any effect anymore. Please
replace them with custom tables and custom filtersets. Also the the
locations and names and structure of the templates changed, so you will
have to update those.