refactor(apis_relations): use generic views as base for views
This commit removes the GenericRelationDetailView, because it was not
being used and is being replaced by the `generic.views.Detail` view.

It also replaces the GenericRelationView, which shows the list of
Relations, with one that inherits from `apis_core.generic.views.List`.
Thas way we can reuse logic of autodetecting the needed tables and
filtersets. Therefore the `rel_filters` module was replaces by a much
smaller `filtersets` module and the `tables` module was cleaned up and
the tables now inherit from `apis_core.generic.tables.GenericTable`

The only thing not touched is the `get_generic_triple_table` method in
the tables, which is used to create the relations table in the entity
views.