feat: add a duplicate button to the generic table
This commit introduces a `duplicate` view in the generic app, which
simply calls the `duplicate()` method of the requested object and
redirects to the `Update` view of the created object.
We also introduce a `duplicate` button in the listview table, that is
only shown if the user has the permission to create the listed object
and if the object has the `duplicate()` method.

Closes: #586