fix: let the autocomplete form create objects
This commit moves the logic of creating a model instance from an URI to
`apis_core.utils.helpers.create_object_from_uri`. This way it can be
reused - and this is exactly what we do in the
`apis_core.generic.views.Autocomplete` view. This view can now create
objects by receiving a POST with a value - it uses
`create_object_from_uri` to do this.

The ModelImportChoiceField now also uses this method to create the
object.