fix(apis_relations): don't rely on APIS_BASE_URI for editing relations
The form for editing relations loads the form and sets the
`other_entity` value to the URI of the other entity. For that it checked
if the URI startswith APIS_BASE_URI - which does not work if the
APIS_BASE_URI is not set. But it doesn't make any sense to make that
`startwith` check. Any URI in the list of URIs should work.