feat(generic): fix and enhance the generate_search_filter helper
The `generate_search_filter` did not split the `query` argument and so
it used every character of the querystring as a lookup query for the
query. This commit uses `query.split()` to split the querystring at
whitespaces and make a list out of it.
In addition to that, it now also allows to pass a list of fields the
search filter should use for the search and only falls back to the
default (of using all charfields) if no such list is provided.