feat(generic): use template attribute of autocomplete view
The Select2QuerySetView can use a template to render autocomplete
results. We use the `get_template_names` method to find a list of
possible template names and if `select_template` finds a valid one, we
use that to render the autocomplete results.

This also adds relevant documentation to the `customzation` chapter.