refactor(utils): replace ASK approach in rdfimport
The approach of our rdf import method, to use an ASK query to check if
an rdfgraph fits an rdfimport definition, turned out to be to limited.
Therefore we replace the approach and go back to the original approach
of defining a target class and a regex. But instead of using a fixed
Python class as a target, we set a parent class and the definition works
for all the subclasses.