feat(generic): allow apiview to use custom serializers & renderers
The commit configures the generic API view to look for more differently
configured serializers. The serializers can now also be prefixed with
the requested format, as configured in the renderer. Additionaly it is
possible to pass a `dialect` queryparameter and the API view now also
uses that to lookup serializers. So in effect, the serializer is looked
up by combining the model, the format and the dialect.
Example: model Person, renderer format: rdf+xml, dialiect: wikidata:
In this order:
* apis_ontology.serializers.PersonWikidataRdfXmlSerializer
* apis_core.apis_entities.serializers.AbstractEntityWikidataRdfXmlSerializer
* apis_core.apis_metainfo.serializers.RootObjectWikidataRdfXmlSerializer
* apis_core.generic.serializers.GenericModelWikidataRdfXmlSerializer
* apis_ontology.serializers.PersonWikidataSerializer
* apis_core.apis_entities.serializers.AbstractEntityWikidataSerializer
* apis_core.apis_metainfo.serializers.RootObjectWikidataSerializer
* apis_core.generic.serializers.GenericModelWikidataSerializer
* apis_ontology.serializers.PersonRdfXmlSerializer
* apis_core.apis_entities.serializers.AbstractEntityRdfXmlSerializer
* apis_core.apis_metainfo.serializers.RootObjectRdfXmlSerializer
* apis_core.generic.serializers.GenericModelRdfXmlSerializer
* apis_ontology.serializers.PersonSerializer
* apis_core.apis_entities.serializers.AbstractEntitySerializer
* apis_core.apis_metainfo.serializers.RootObjectSerializer
* apis_core.generic.serializers.GenericModelSerializer'