refactor(utils,default_settings): remove unused modules
The `utils.stanbolQueries` was not being used anywhere in the code and
we therefore remove it. This file and the `apis_entities.autocomplete3`
module were the only ones that referenced the `NER_settings` module,
though it wasn't actually used in the `autocomplete3`. The reference to
the `NER_settings` was therefore also removed and in turn the
`NER_settings` itself, too.