refactor(apis_metainfo,apis_entities)!: drop __str___ overrides
The `RootObject` and the `AbstractEntity` models had an override for the
`__str__` method, that referenced the models `.name` field. We are
planning on removing the `.name` field and the `__str__` methods should
be implemented in child classes anyway, so this commit removes the
overrides from `RootObject` and `AbstractEntity`
It also deletes the test, as we are now using Django's default
behaviour.

cherry picked from apis-core-rdf#630 by birger