refactor(apis_metainfo)!: rename RootObject.name to deprecated_name
This commit renames the RootObject's `.name` field to
`.deprecated_name`. This is a preparation for dropping the field
altogehter later.
The field is now also allowed to be empty.

This is a cherry-pick of the chanes in apis-core-rdf#630

BREAKING-CHANGE: The renaming of RootObject's "name" field to
"deprecated_name" affects all inheriting model classes as well as class
Property. You will need to create new fields in your model classes to
replace it/to preserve its contents before it gets dropped.

Thanks to KK for doing all the research and propose the changes!

Closes: #299