refactor(utils): use contenttype instead of model name for helper
The `get_classes_with_allowed_relation_from` method created a list of
possible related models based on the name of one model. It then returned
the names of those models as strings. It makes more sense to simply use
contenttypes instead of model names, especially since the Property model
stores the possible subj and obj classes as ContentTypes.
This commit refactors the method and also updates `triple_sidebar` which
is the only funciton that uses the
`get_classes_with_allowed_relation_from` method.

Closes: #254