fix(generic): use dal widget only for models with autocomplete endpoint
The generic form overrides the default select widgets with autocomplete
widgets provided by the django autocomplete light app. This can only
work for contenttypes that provide an autocomplete endpoint. We can only
be sure of that for models that inherit from the GenericModel, so we
skip the override for all the others.

Closes: #764