fix(history): use signal instead of method for setting timestamp
When doing a CASCADING delete, the `delete` method of an instance is not
called, but the `pre_delete` signal is triggered. We want to set the
current date for the `_history_date` attribute, if it is None. To do
that also in cascading deletes, we use the pre_delete signal.

Closes: #814