fix(apis_entities): remove None values before checking model_class
In rare cases (i.e. when you remove a model and don't cleanup the
ContentTypes) we end up having ContentTypes in Djangos DB that have
None as a `model_class`. But we can not check if None is a subclass, so
we have to remove those instance before doing the check.