fix(apis_relations,apis_entities): replace get_object_or_404
`get_object_or_404` errors if we are filtering for the modelname only
(before, we filtered for modelname in the `apis_ontology` app). When we
work with models that have the same name as models in Django (like
`Group`) this leads to multiple results, which `get_object_or_404` does
not like. Therefore we are now first making a list of contenttypes that
are entities, then we filter for the model name, then we pick the first
one.
This whole nonsense will be replaced anyway, when we implement new
relations...