feat(generic): list all generic models on overview page
This commit introduces the templatetag `genericmodel_content_types`
which lists all content types of models that inherit from
`apis_core.generic.abc.GenericModel`. This templatetag is then used
in the overview template to create an overview of *all* those
contenttypes and not only `apis_ontology` and `collections` as it was
hardcoded before.

Closes: #912