feat(generic): add CustomAPIRootView & CustomDefaultRouter
The DRF default APIRootView that is used by the DRF DefaultRouter only
lists routes that have been registered using `router.register` without a
parameter as part of the URL. We want it to also list routes that we
registered with `<content_type>` as URI part. Therefore we have to
inject those routes manually, which we do using a CustomAPIRootView and
a CustomDefaultRouter.

Closes: #836