fix: deactivate simple_history plugin for loaddata
resolves #915
In the fixtures m2m fields are populated directly when the object
is created. Given that we activate the tracking of m2m fields by
default in simple_history loaddata creates not null constraints
errors for objects that have m2m relations. Also the history is anyway
in the fixtures, so no need to activate the plugin during import of
the data.