refactor(apis_relations): drop use of filter and filter itself
The `content_type` filter was only used in the bibsonomy button
template. This template did not work anymore the way it was set up, so
it is being replaced and the filter is dropped.
The filter was the only one in the `apis_helpers` templatetags, so we
delete the whole templatetag file.