refactor(utils)!: drop `filtermethods`
The `utils.filtermethods` module contained a couple of unused
filtermethods that were used in old APIS models and referenced hardcoded
field names. They were all replaced by more flexible approaches by now.
The `construct_lookup` method that was part of the module can be useful
and was therefore salvaged and moved to the `utils.helpers` module.
The `test_filtermethods` module was renamed accordingly.

Closes: #790