fix(generic): repair table columns ordering
We interpreted a GET request *with* parameters but *without* a columns
parameter as the wish to have a table without any columns. But that
could also be a wish to order the table, which only appends a `sort`
parameter to existing parameters (or creates parameters if there are
none).
Therefore for now we fallback to the default columns if none are
selected, but we gain table ordering.

Closes: #1022