feat(apis_entites): add a generic entites autocomplete endpoint
This endpoint allows us to use autocomplete over multiple model types.
It takes a parameter `entities` which is a list of ContentType natural
keys and searches for the query in all instances of those entities
(using `generate_search_filter`, which means it uses a different search
approach for every model).
The return values of the endpoint are then prefixed with the id of the
contenttype of the results, separated by an underscore.