fix(relations): make `get_relation_targets_from` method more targeted
The `get_relation_targets_from` method listed all possible targets of
all the relations an object might be related to. But we actually want
only a list of the targets of the relations *other* sides, seen from the
objects point of view. This is why we now also instead pass the object to
`get_relation_targets_from` and use that to check for the right targets.