feat(relations): extend the base template to include relation menu items
This introduces a `base.html` template in the relations template folder,
which extends the default `base.html` and adds menu items to the
relations menu. It adds one item pointing to the overview of all the
existing relations as well as one menu item for each relation type,
pointing to the listview of that relation type. The template also
includes a javascript file that allows to search the relation menu
items if there are more than 10.

Important: for this to work, `apis_core.relations` has to come before
`apis_core.core` in the INSTALLED_APPS.

Closes: #1083

Co-authored-by: Matthias Schlögl <matthias.schloegl@oeaw.ac.at>