fix(apis_entities): use example.org as fallback and drop global
Use the name `example.org` as fallback for the base domain of new URIs,
if there is no APIS_BASE_URI set in the settings. example.org is a
domain for exactly that purpose, to provide an example. See also
https://www.iana.org/help/example-domains
(apis.info on the other hand points to a commercial domain registry and
costs ~10.000$)
To make things easier to read, the useless use of a global was removed.