feat(generic): add a view for enriching entities with external data
This view uses the importer to merge data from an URI (which is passed
as a GET parameter). If the URI is connected to another model instance,
the view redirects to the merge view. Otherwise the data from the URI
is fetched and presented to the user as a list of checkboxes for the
values that should be copied into the model instance.
If the URI is not yet connected to the instance, a new Uri
instance will be created and connected to the model instance.