fix(apis_metainfo): drop UriGetOrCreate view and related stuff
The UriGetOrCreate view used the UriGetOrCreateForm, which did not what
the name suggested: it did get_or_create an Uri, but if the created uri
instance did not link to a root_object it rightaway deleted the it and
raised a ValidationError. This basically boils down to a simple `get`.

This also removes the templates used by the view, the mentioned form, and
the whole urls module, which only pointed to the view.

Closes: #1135